\name{YEASTREJECTORF}
\alias{YEASTREJECTORF}
\title{Real and Rejected Yeast Genes (ORF)}
\description{
  Real and rejected yeast ORFs from Kellis et al. (2003)
}
\details{
  A character vector which contains yeast ORFs which are rejected
  in the reading frame conseration (RFC) test in Kellis et al. (2003).
}
\references{
  Manolis Kellis, Nick Patterson, Matthew Endrizzi, Bruce Birren and
  Eric S. Lander, Sequencing and comparison of yeast species to identify
  genes and regulatory elements.  Nature 423, 241-254 (15 May 2003)
}
\examples{
library(YEAST)
head(YEASTREJECTORF)
}
\keyword{datasets}

